+++ common/src/main/java/net/frozenblock/lib/entity/api/command/ScaleEntityCommand.java
@@ -70,5 +70,4 @@ public class ScaleEntityCommand {
 			return 0;
 		}
 	}
-
 }