+++ common/src/main/java/net/frozenblock/lib/levelgen/structure/impl/status/StructureStatus.java
@@ -45,7 +45,7 @@ public record StructureStatus(Identifier structure, boolean insidePiece) {
 
 	public static Optional<StructureStatus> getProminentStructureStatus(Player player) {
 		if (player == null) return Optional.empty();
-		final List<StructureStatus> statuses = ATTACHMENT_TYPE.getAttachedOrGet(player, ImmutableList::of);
+		final List<StructureStatus> statuses = ATTACHMENT_TYPE.getAttachedOrElseGet(player, ImmutableList::of);
 		return Optional.ofNullable(statuses.stream()
 			.filter(StructureStatus::insidePiece)
 			.findFirst()